+++ pkg/icingadb/cleanup.go
@@ -5,7 +5,6 @@ import (
 	"fmt"
 	"github.com/icinga/icingadb/internal"
 	"github.com/icinga/icingadb/pkg/com"
-	"github.com/icinga/icingadb/pkg/driver"
 	"github.com/icinga/icingadb/pkg/types"
 	"time"
 )
@@ -20,10 +19,10 @@ type CleanupStmt struct {
 // Build assembles the cleanup statement for the specified database driver with the given limit.
 func (stmt *CleanupStmt) Build(driverName string, limit uint64) string {
 	switch driverName {
-	case driver.MySQL:
+	case MySQL:
 		return fmt.Sprintf(`DELETE FROM %[1]s WHERE environment_id = :environment_id AND %[2]s < :time
 ORDER BY %[2]s LIMIT %[3]d`, stmt.Table, stmt.Column, limit)
-	case driver.PostgreSQL:
+	case PostgreSQL:
 		return fmt.Sprintf(`WITH rows AS (
 SELECT %[1]s FROM %[2]s WHERE environment_id = :environment_id AND %[3]s < :time ORDER BY %[3]s LIMIT %[4]d
 )
