+++ pkg/icingadb/db.go
@@ -7,7 +7,6 @@ import (
 	"github.com/icinga/icingadb/pkg/backoff"
 	"github.com/icinga/icingadb/pkg/com"
 	"github.com/icinga/icingadb/pkg/contracts"
-	"github.com/icinga/icingadb/pkg/driver"
 	"github.com/icinga/icingadb/pkg/logging"
 	"github.com/icinga/icingadb/pkg/periodic"
 	"github.com/icinga/icingadb/pkg/retry"
@@ -97,9 +96,9 @@ const (
 func (db *DB) CheckSchema(ctx context.Context) error {
 	var expectedDbSchemaVersion uint16
 	switch db.DriverName() {
-	case driver.MySQL:
+	case MySQL:
 		expectedDbSchemaVersion = expectedMysqlSchemaVersion
-	case driver.PostgreSQL:
+	case PostgreSQL:
 		expectedDbSchemaVersion = expectedPostgresSchemaVersion
 	}
 
@@ -165,10 +164,10 @@ func (db *DB) BuildInsertIgnoreStmt(into interface{}) (string, int) {
 	var clause string
 
 	switch db.DriverName() {
-	case driver.MySQL:
+	case MySQL:
 		// MySQL treats UPDATE id = id as a no-op.
 		clause = fmt.Sprintf(`ON DUPLICATE KEY UPDATE "%s" = "%s"`, columns[0], columns[0])
-	case driver.PostgreSQL:
+	case PostgreSQL:
 		clause = fmt.Sprintf("ON CONFLICT ON CONSTRAINT pk_%s DO NOTHING", table)
 	}
 
@@ -228,10 +227,10 @@ func (db *DB) BuildUpsertStmt(subject interface{}) (stmt string, placeholders in
 
 	var clause, setFormat string
 	switch db.DriverName() {
-	case driver.MySQL:
+	case MySQL:
 		clause = "ON DUPLICATE KEY UPDATE"
 		setFormat = `"%[1]s" = VALUES("%[1]s")`
-	case driver.PostgreSQL:
+	case PostgreSQL:
 		clause = fmt.Sprintf("ON CONFLICT ON CONSTRAINT pk_%s DO UPDATE SET", table)
 		setFormat = `"%[1]s" = EXCLUDED."%[1]s"`
 	}
